Project Forms for Staff and related entities
We have each uploaded the forms corresponding to the ERD we presented earlier. They are not linked to one another as of now due to time constrainsts because the login form will redirect the user according to the category he/she belongs to. This can only be done once we have the IDs and user names segregated according to student,faculty and staff.